refactor(preset-mini,preset-wind): move variants to preset-wind
- move `dark`/`light` variants
- move `motion-reduce`/`motion-safe`
- move `portrait`/`landscape`
