+++ packages/preset-mini/src/variants/default.ts
@@ -3,10 +3,9 @@ import type { PresetMiniOptions } from '..'
 import type { Theme } from '../theme'
 import { variantBreakpoints } from './breakpoints'
 import { variantCombinators } from './combinators'
-import { variantColorsMediaOrClass } from './dark'
 import { variantLanguageDirections } from './directions'
 import { variantImportant, variantLayer, variantNegative, variantScope } from './misc'
-import { variantCustomMedia, variantMotions, variantOrientations, variantPrint } from './media'
+import { variantCustomMedia, variantPrint } from './media'
 import { partClasses, variantPseudoClassFunctions, variantPseudoClasses, variantPseudoElements, variantTaggedPseudoClasses } from './pseudo'
 
 export const variants = (options: PresetMiniOptions): Variant<Theme>[] => [
@@ -15,8 +14,6 @@ export const variants = (options: PresetMiniOptions): Variant<Theme>[] => [
   variantImportant,
   variantPrint,
   variantCustomMedia,
-  ...variantOrientations,
-  ...variantMotions,
   variantBreakpoints,
   ...variantCombinators,
   variantPseudoClasses,
@@ -24,7 +21,6 @@ export const variants = (options: PresetMiniOptions): Variant<Theme>[] => [
   ...variantTaggedPseudoClasses(options),
   variantPseudoElements,
   partClasses,
-  ...variantColorsMediaOrClass(options),
   ...variantLanguageDirections,
   variantScope,
 ]
