+++ packages/preset-mini/src/variants/index.ts
@@ -2,7 +2,6 @@
 export * from './breakpoints'
 export * from './combinators'
 export * from './media'
-export * from './dark'
 export * from './default'
 export * from './directions'
 export * from './misc'
