+++ packages/preset-wind/src/index.ts
@@ -11,7 +11,14 @@ export type { Theme } from '@unocss/preset-mini'
 
 export { rules, shortcuts, theme, variants }
 
-export interface UnoOptions extends PresetMiniOptions { }
+interface DarkOption {
+  /**
+   * @default 'class'
+   */
+  dark?: 'class' | 'media'
+}
+
+export interface UnoOptions extends DarkOption, PresetMiniOptions { }
 
 export const presetWind = (options: UnoOptions = {}): Preset<Theme> => {
   options.dark = options.dark ?? 'class'
