+++ packages/preset-wind/src/variants/default.ts
@@ -2,7 +2,8 @@ import type { Variant } from '@unocss/core'
 import { variants as miniVariants } from '@unocss/preset-mini/variants'
 import type { Theme, UnoOptions } from '..'
 import { variantCombinators } from './combinators'
-import { variantColorsScheme } from './dark'
+import { variantColorsMediaOrClass, variantColorsScheme } from './dark'
+import { variantMotions, variantOrientations } from './media'
 import { variantSpaceAndDivide } from './misc'
 import { placeholderModifier } from './placeholder'
 
@@ -10,6 +11,9 @@ export const variants = (options: UnoOptions): Variant<Theme>[] => [
   placeholderModifier,
   variantSpaceAndDivide,
   ...miniVariants(options),
+  ...variantOrientations,
+  ...variantMotions,
   ...variantCombinators,
   ...variantColorsScheme,
+  ...variantColorsMediaOrClass(options),
 ]
