+++ test/preset-mini-targets.ts
@@ -651,11 +651,6 @@ export const presetMiniTargets: string[] = [
   '[content:attr(attr\\_content)]',
   '[background-image:url(star_transparent.gif),_url(cat_front.png)]',
 
-  // dark/light
-  'dark:not-odd:text-red',
-  'dark:text-xl',
-  'light:text-sm',
-
   // variants
   'active:scale-4',
   'hover:scale-4',
@@ -715,14 +710,6 @@ export const presetMiniTargets: string[] = [
 
   // variants media
   'media-cssvar:block',
-
-  // variants motion
-  'motion-reduce:hover:translate-0',
-  'motion-safe:transition',
-
-  // variants orientations
-  'portrait:hidden',
-  'landscape:hidden',
 
   // variants prints
   'print:block',
